requirements.txt
Flask==2.1.3
gunicorn==20.1.0
celery==5.3.5
redis==4.3.5
requests==2.28.1
Flask-Mail==0.9.0
Flask-JWT-Extended==4.4.4
Flask-PyMongo==2.3.0
Werkzeug==2.2.2
Flask-PyMongo==2.3.0
flasgger==0.9.7.1
Flask-Cors==3.0.10
coverage==7.1.0
pytest==7.2.1
sentry-sdk==1.4.3
pycryptodome==3.18.0
cryptography==41.0.7
python-dotenv==1.0.0
rsa==4.9
pytest-dotenv==0.5.2
flask-babel==4.0.0
Flask-WTF==1.2.1
pydantic==2.11.3
pydantic_core==2.33.1
watchdog==6.0.0
